Separated all different section to their own component and added some new styling with new projects
