+++ package.json
@@ -21,6 +21,7 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-icons": "^5.5.0",
+    "react-simple-typewriter": "^5.0.1",
     "tailwind-merge": "^3.3.1",
     "tailwindcss": "^4.1.11"
   },
